Move experimental stock metrics to own .py for readability
